--- Lab1/Text/1.txt
@@ -1,2 +1,0 @@
-To explore strange new worlds,
-To seek out new life and new civilizations
